Initial drop of composite simulation code.
Composite simulations bundle other simulations; simulations are run in steps, and connected via a mapping between variables.  The execution of a composite simulation results in a composite scenario that embeds partial results from each step.  This makes it possible to re-run a single step in a simulation if we wish to, for instance, inject a new model as a component of a composite simulation.