Lots of refactoring in order to support client side Proxy object generation.
This was a long and winding path through hell. Roo should not be used any longer.  Upgraded aspectj plugin to recommended 1.3.1 to address weaving problems.  Substantive changes include the addition of a "Variable" interface, and a user collection type to deal with a JAXB issue.  This is generally speaking a step *backwards* due to unforeseen complications with technology dependencies.